+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 Changelog
 =========
 
+Version 0.1.1.0
+---------------
+
+*Unreleased*
+
+<https://github.com/mstksg/decidable/releases/tag/v0.1.1.0>
+
+*   `flipDecision`, `forgetDisproof`, `forgetProof`, `isProved`, and
+    `isDisproved` added to *Data.Type.Predicate* module.
+
 Version 0.1.0.0
 ---------------